Moved login files to components
